requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
pdf2image==1.17.0
gunicorn==21.2.0
Pillow==10.0.0
Flask-Login==0.6.3
pdf2docx==0.5.6
